Slight model change, and the beginnings of the post job view
One attribute was changed from an int to a double in the model. The post
job view is nearing completion (needs some more functionality, and some
component positioning tweaks), but an interesting bug has come up. For
some reason the app won't save the LoggedIn user. It saves it in the
unit test, but for some reason it is being difficult in the app itself.
It's likely a small bug with the controller, but I'm having trouble
finding it right now, so I'm going to leave it until tomorrow.

+++ DesktopTAMAS/src/ecse321/group12/tamas/model/Course.java
@@ -381,7 +381,7 @@ public class Course
     return 0;
   }
 
-  public Job addJob(int aMaxHours, int aWage, Date aDeadline, String aRequiredSkills, String aRequiredCourseGPA, String aRequiredCGPA, String aRequiredExperience)
+  public Job addJob(int aMaxHours, double aWage, Date aDeadline, String aRequiredSkills, String aRequiredCourseGPA, String aRequiredCGPA, String aRequiredExperience)
   {
     return new Job(aMaxHours, aWage, aDeadline, aRequiredSkills, aRequiredCourseGPA, aRequiredCGPA, aRequiredExperience, this);
   }
