+++ DesktopTAMAS/src/ecse321/group12/tamas/model/TAjob.java
@@ -21,7 +21,7 @@ public class TAjob extends Job
   // CONSTRUCTOR
   //------------------------
 
-  public TAjob(int aMaxHours, int aWage, Date aDeadline, String aRequiredSkills, String aRequiredCourseGPA, String aRequiredCGPA, String aRequiredExperience, Course aCourse, int aMinHours, boolean aIsLab)
+  public TAjob(int aMaxHours, double aWage, Date aDeadline, String aRequiredSkills, String aRequiredCourseGPA, String aRequiredCGPA, String aRequiredExperience, Course aCourse, int aMinHours, boolean aIsLab)
   {
     super(aMaxHours, aWage, aDeadline, aRequiredSkills, aRequiredCourseGPA, aRequiredCGPA, aRequiredExperience, aCourse);
     minHours = aMinHours;
